Move LayoutEditorDialog to baaahs.app.ui.editor.layout.
